Refactor: now an arithmetic op
