+++ pennylane/ops/qubit/arithmetic_ops.py
@@ -16,10 +16,15 @@ This submodule contains the discrete-variable quantum operations that perform
 arithmetic operations on their input states.
 """
 # pylint:disable=abstract-method,arguments-differ,protected-access
+from copy import copy
+
 import numpy as np
 
 import pennylane as qml
-from pennylane.operation import Operation
+from pennylane.operation import AnyWires, Operation
+from pennylane.wires import Wires
+
+from .non_parametric_ops import MultiControlledX
 
 
 class QubitCarry(Operation):
@@ -305,3 +310,178 @@ class QubitSum(Operation):
 
     def adjoint(self):
         return QubitSum(wires=self.wires)
+
+
+class IntegerComparator(Operation):
+    r"""IntegerComparator(value, geq, wires)
+    Apply a controlled Pauli X gate using integer comparison as the condition.
+
+    Given a basis state :math:`\vert n \rangle`, where :math:`n` is a positive integer, and a fixed positive
+    integer :math:`L`, flip a target qubit if :math:`n \geq L`. Alternatively, the flipping condition can
+    be :math:`n < L`.
+
+    **Details:**
+
+    * Number of wires: Any (the operation can act on any number of wires)
+    * Number of parameters: 1
+    * Gradient recipe: None
+
+    .. note::
+
+        This operation has one parameter: ``value``. However, ``value`` is simply an integer that is required to define
+        the condition upon which a Pauli X gate is applied to the target wire. Given that, IntegerComparator has a
+        gradient of zero; ``value`` is a non-differentiable parameter.
+
+    Args:
+        value (int): The value :math:`L` that the state's decimal representation is compared against.
+        geq (bool): If set to ``True``, the comparison made will be :math:`\sigma \geq L`. If ``False``, the comparison
+            made will be :math:`\sigma < L`.
+        wires (Union[Wires, Sequence[int], or int]): Control wire(s) followed by a single target wire where
+            the operation acts on.
+
+    **Example**
+
+    dev = qml.device("default.qubit", wires=3)
+
+    @qml.qnode(dev)
+    def circuit(state, value, geq):
+        qml.BasisState(np.array(state), wires=range(3))
+        qml.IntegerComparator(value, geq=geq, wires=range(3))
+        return qml.state()
+
+    >>> circuit([1, 0, 1], 1, True).reshape(2, 2, 2)[1, 0, 0]
+    (1+0j)
+    >>> circuit([0, 1, 0], 3, False).reshape(2, 2, 2)[0, 1, 1]
+    (1+0j)
+    """
+    is_self_inverse = True
+    num_wires = AnyWires
+    num_params = 0
+    """int: Number of trainable parameters that the operator depends on."""
+
+    grad_method = None
+    # pylint: disable=too-many-arguments
+    def __init__(
+        self,
+        value,
+        geq=True,
+        wires=None,
+        do_queue=True,
+    ):
+
+        if not isinstance(value, int):
+            raise ValueError("The comparable value must be an integer.")
+        if wires is None:
+            raise ValueError("Must specify the target wire where the operation acts on.")
+        if len(wires) > 1:
+            control_wires = Wires(wires[:-1])
+            wires = Wires(wires[-1])
+        else:
+            raise ValueError(
+                "IntegerComparator: wrong number of wires. "
+                f"{len(wires)} wire(s) given. Need at least 2."
+            )
+
+        total_wires = control_wires + wires
+
+        self.hyperparameters["control_wires"] = control_wires
+        self.hyperparameters["target_wires"] = wires
+        self.hyperparameters["value"] = value
+        self.hyperparameters["geq"] = geq
+        self.value = value
+
+        super().__init__(wires=total_wires, do_queue=do_queue)
+
+    def label(self, decimals=None, base_label=None, cache=None):
+        return base_label or f">={self.value}" if self.geq else f"<{self.value}"
+
+    # pylint: disable=unused-argument
+    @staticmethod
+    def compute_matrix(
+        value=None, control_wires=None, geq=True, **kwargs
+    ):  # pylint: disable=arguments-differ
+        r"""Representation of the operator as a canonical matrix in the computational basis (static method).
+
+        The canonical matrix is the textbook matrix representation that does not consider wires.
+        Implicitly, this assumes that the wires of the operator correspond to the global wire order.
+
+        .. seealso:: :meth:`~.IntegerComparator.matrix`
+
+        Args:
+            value (int): The value :math:`L` that the state's decimal representation is compared against.
+            control_wires (Union[Wires, Sequence[int], or int]): wires to place controls on
+            geq (bool): If set to `True`, the comparison made will be :math:`n \geq L`. If `False`, the comparison
+                made will be :math:`n < L`.
+
+        Returns:
+           tensor_like: matrix representation
+
+        **Example**
+
+        >>> print(qml.IntegerComparator.compute_matrix(2, [0,1]))
+        [[1. 0. 0. 0. 0. 0. 0. 0.]
+         [0. 1. 0. 0. 0. 0. 0. 0.]
+         [0. 0. 1. 0. 0. 0. 0. 0.]
+         [0. 0. 0. 1. 0. 0. 0. 0.]
+         [0. 0. 0. 0. 0. 1. 0. 0.]
+         [0. 0. 0. 0. 1. 0. 0. 0.]
+         [0. 0. 0. 0. 0. 0. 0. 1.]
+         [0. 0. 0. 0. 0. 0. 1. 0.]]
+        >>> print(qml.IntegerComparator.compute_matrix(2, [0,1], geq=False))
+        [[0. 1. 0. 0. 0. 0. 0. 0.]
+         [1. 0. 0. 0. 0. 0. 0. 0.]
+         [0. 0. 0. 1. 0. 0. 0. 0.]
+         [0. 0. 1. 0. 0. 0. 0. 0.]
+         [0. 0. 0. 0. 1. 0. 0. 0.]
+         [0. 0. 0. 0. 0. 1. 0. 0.]
+         [0. 0. 0. 0. 0. 0. 1. 0.]
+         [0. 0. 0. 0. 0. 0. 0. 1.]]
+        """
+
+        if value is None:
+            raise ValueError("The value to compare to must be specified.")
+        if control_wires is None:
+            raise ValueError("Must specify the control wires.")
+        if isinstance(value, int):
+            binary = "0" + str(len(control_wires)) + "b"
+
+            if geq:
+                if value > 2 ** len(control_wires) - 1:
+                    mat = np.eye(2 ** (len(control_wires) + 1))
+
+                else:
+                    control_values_list = [
+                        format(n, binary) for n in range(value, 2 ** (len(control_wires)))
+                    ]
+                    mat = np.eye(2 ** (len(control_wires) + 1))
+                    for control_values in control_values_list:
+                        mat = mat @ MultiControlledX.compute_matrix(
+                            control_wires, control_values=control_values
+                        )
+
+            else:
+                if value == 0:
+                    mat = np.eye(2 ** (len(control_wires) + 1))
+
+                else:
+                    control_values_list = [format(n, binary) for n in range(value)]
+                    mat = np.eye(2 ** (len(control_wires) + 1))
+                    for control_values in control_values_list:
+                        mat = mat @ MultiControlledX.compute_matrix(
+                            control_wires, control_values=control_values
+                        )
+
+        else:
+            raise ValueError(f"The compared value must be an int. Got {type(value)}.")
+
+        return mat
+
+    @property
+    def control_wires(self):
+        return self.wires[:~0]
+
+    def adjoint(self):
+        return copy(self).queue()
+
+    def pow(self, z):
+        return super().pow(z % 2)
