+++ pennylane/ops/qubit/non_parametric_ops.py
@@ -2243,170 +2243,6 @@ class MultiControlledX(Operation):
         return True
 
 
-class IntegerComparator(Operation):
-    r"""IntegerComparator(value, geq, wires)
-    Apply a controlled Pauli X gate using integer comparison as the condition.
-
-    Given a basis state :math:`\vert \sigma \rangle`, where :math:`\sigma \in \mathbb{N}`, and a fixed positive
-    integer :math:`L`, flip a target qubit if :math:`\sigma \geq L`. Alternatively, the flipping condition can
-    be :math:`\sigma < L`.
-
-    **Details:**
-
-    * Number of wires: Any (the operation can act on any number of wires)
-    * Number of parameters: 0
-    * Gradient recipe: None
-
-    Args:
-        value (int): The value :math:`L` that the state's decimal representation is compared against.
-        geq (bool): If set to ``True``, the comparison made will be :math:`\sigma \geq L`. If ``False``, the comparison
-            made will be :math:`\sigma < L`.
-        wires (Union[Wires, Sequence[int], or int]): Control wire(s) followed by a single target wire where
-            the operation acts on.
-
-    **Example**
-
-    dev = qml.device("default.qubit", wires=3)
-
-    @qml.qnode(dev)
-    def circuit(state, value, geq):
-        qml.BasisState(np.array(state), wires=range(3))
-        qml.IntegerComparator(value, geq=geq, wires=range(3))
-        return qml.state()
-
-    >>> circuit([1, 0, 1], 1, True).reshape(2, 2, 2)[1, 0, 0]
-    (1+0j)
-    >>> circuit([0, 1, 0], 3, False).reshape(2, 2, 2)[0, 1, 1]
-    (1+0j)
-    """
-    is_self_inverse = True
-    num_wires = AnyWires
-    num_params = 0
-    """int: Number of trainable parameters that the operator depends on."""
-
-    grad_method = None
-    # pylint: disable=too-many-arguments
-    def __init__(
-        self,
-        value,
-        geq=True,
-        wires=None,
-        do_queue=True,
-    ):
-
-        if value is None:
-            raise ValueError("Must specify the integer value to compare against.")
-        if not isinstance(value, int):
-            raise ValueError("The comparable value must be an integer.")
-        if wires is None:
-            raise ValueError("Must specify the target wire where the operation acts on.")
-        if len(wires) > 1:
-            control_wires = Wires(wires[:-1])
-            wires = Wires(wires[-1])
-        else:
-            raise ValueError(
-                "IntegerComparator: wrong number of wires. "
-                f"{len(wires)} wire(s) given. Need at least 2."
-            )
-
-        total_wires = control_wires + wires
-
-        self.hyperparameters["control_wires"] = control_wires
-        self.hyperparameters["value"] = value
-        self.hyperparameters["geq"] = geq
-        self.value = value
-
-        super().__init__(wires=total_wires, do_queue=do_queue)
-
-    def label(self, decimals=None, base_label=None, cache=None):
-        return base_label or f">={self.value}" if self.geq else f"<{self.value}"
-
-    # pylint: disable=unused-argument
-    @staticmethod
-    def compute_matrix(
-        value=None, control_wires=None, geq=True, **kwargs
-    ):  # pylint: disable=arguments-differ
-        r"""Representation of the operator as a canonical matrix in the computational basis (static method).
-
-        The canonical matrix is the textbook matrix representation that does not consider wires.
-        Implicitly, this assumes that the wires of the operator correspond to the global wire order.
-
-        .. seealso:: :meth:`~.IntegerComparator.matrix`
-
-        Args:
-            value (int): The value :math:`L` that the state's decimal representation is compared against.
-            control_wires (Union[Wires, Sequence[int], or int]): wires to place controls on
-            geq (bool): If set to `True`, the comparison made will be :math:`\sigma \geq L`. If `False`, the comparison
-                made will be :math:`\sigma < L`.
-
-        Returns:
-           tensor_like: matrix representation
-
-        **Example**
-
-        >>> print(qml.IntegerComparator.compute_matrix(2, [0,1]))
-        [[1. 0. 0. 0. 0. 0. 0. 0.]
-         [0. 1. 0. 0. 0. 0. 0. 0.]
-         [0. 0. 1. 0. 0. 0. 0. 0.]
-         [0. 0. 0. 1. 0. 0. 0. 0.]
-         [0. 0. 0. 0. 0. 1. 0. 0.]
-         [0. 0. 0. 0. 1. 0. 0. 0.]
-         [0. 0. 0. 0. 0. 0. 0. 1.]
-         [0. 0. 0. 0. 0. 0. 1. 0.]]
-        >>> print(qml.IntegerComparator.compute_matrix(2, [0,1], geq=False))
-        [[0. 1. 0. 0. 0. 0. 0. 0.]
-         [1. 0. 0. 0. 0. 0. 0. 0.]
-         [0. 0. 0. 1. 0. 0. 0. 0.]
-         [0. 0. 1. 0. 0. 0. 0. 0.]
-         [0. 0. 0. 0. 1. 0. 0. 0.]
-         [0. 0. 0. 0. 0. 1. 0. 0.]
-         [0. 0. 0. 0. 0. 0. 1. 0.]
-         [0. 0. 0. 0. 0. 0. 0. 1.]]
-        """
-
-        if value is None:
-            raise ValueError("The value to compare to must be specified.")
-        if control_wires is None:
-            raise ValueError("Must specify the control wires.")
-        if isinstance(value, int):
-            paulix_mat = qml.PauliX.compute_matrix()
-
-            if geq:  # comparison is i >= value
-                if value > 2 ** len(control_wires) - 1:  # if value exceeds Hilbert space size
-                    mat = np.eye(2 ** (len(control_wires) + 1))
-
-                else:
-                    padding_left = value * 2
-                    paulix_padding = (2 ** (len(control_wires) + 1) - padding_left) // 2
-                    paulix_blocks = [paulix_mat for _ in range(paulix_padding)]
-                    mat = block_diag(np.eye(padding_left), *paulix_blocks)
-
-            else:  # comparison is i < value
-                if value == 0:
-                    mat = np.eye(2 ** (len(control_wires) + 1))
-
-                else:
-                    padding_right = (2 ** (len(control_wires)) - value) * 2
-                    paulix_padding = (2 ** (len(control_wires) + 1) - padding_right) // 2
-                    paulix_blocks = [paulix_mat for _ in range(paulix_padding)]
-                    mat = block_diag(*paulix_blocks, np.eye(padding_right))
-
-        else:
-            raise ValueError(f"The compared value must be an int. Got {type(value)}.")
-
-        return mat
-
-    @property
-    def control_wires(self):
-        return self.wires[:~0]
-
-    def adjoint(self):
-        return copy(self).queue()
-
-    def pow(self, z):
-        return super().pow(z % 2)
-
-
 class Barrier(Operation):
     r"""Barrier(wires)
     The Barrier operator, used to separate the compilation process into blocks or as a visual tool.
